# backend/api.yaml
openapi: 3.0.3
info:
  title: Cumulus dashboard
  description: |-
    OpenAPI specification for the Cumulus dashboard backend

    Some useful links:
    - [GitHub Repository](https://github.com/smart-on-fhir/cumulus-app)
    - [Test Instance](https://smart-cumulus-test.herokuapp.com/)

  # termsOfService: http://swagger.io/terms/
  # contact:
  #  email: apiteam@swagger.io
  license:
   name: Apache 2.0
   url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
# externalDocs:
#   description: Find out more about Swagger
#   url: http://swagger.io
servers:
  - url: https://smart-cumulus-test.herokuapp.com/api
  - url: https://smart-cumulus-ucdavis.herokuapp.com/api
  - url: https://smart-cumulus-regi-e973987fb5b9.herokuapp.com/api
  - url: https://smart-cumulus-staging.herokuapp.com/api
  - url: https://smart-cumulus.herokuapp.com/api
security: 
- session_id: []
paths:
...
  /views:
    get:
      summary: Get All Views
      description: Retrieve an array of all the views that the current user 
        is allowed to see
      operationId: getViews
      tags: 
        - Views
      parameters:
        - $ref: '#/components/parameters/order'
        - name: limit
          in: query
          schema:
            type: integer
          example: 10
          required: false
          allowEmptyValue: true
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/attributes'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/View'
                    - type: object
                      properties:
                        Tags:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                $ref: '#/components/schemas/standardID'
                              name:
                                type: string
                                maxLength: 50
                              description:
                                type: string
                                maxLength: 200
                        DataRequest:
                          type: object
                          properties:
                            id:
                              $ref: '#/components/schemas/standardID'
                            name:
                              type: string
                              maxLength: 100
              example:
                - id: 1
                  name: Example View
                  description: Example View Description
                  screenShot: Base 64 image data...
                  settings: {}
                  subscriptionId: 1
                  creatorId: 2
                  createdAt: 2024-02-22T13:36:17.315Z
                  updatedAt: 2024-02-22T13:36:17.315Z
                  Tags: []
                  DataRequest:
                    id: 1
                    name: Subscription Name
        '401':
          description: If the user is not authorized and guest users are not
            allowed to view included resources
        '403':
          description: If the user is not allowed to view included resources
        '400':
          description: Bad request due to invalid parameters
    post:
      summary: Create View
      description: Creates new View
      operationId: createView
      tags: 
        - Views
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                  example: View display name
                description:
                  type: string
                  maxLength: 500
                  example: View description
                screenShot:
                  type: string
                  example: base64imageData...
                Tags:
                  type: array
                  items:
                    $ref: '#/components/schemas/Tag'
                settings:
                  type: object
                subscriptionId:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/View'
                  - type: object
                    properties:
                      Tags:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              $ref: '#/components/schemas/standardID'
                            name:
                              type: string
                              maxLength: 50
                            description:
                              type: string
                              maxLength: 200
                      DataRequest:
                        type: object
                        properties:
                          id:
                            $ref: '#/components/schemas/standardID'
                          name:
                            type: string
                            maxLength: 100
        '401':
          description: If the user is not authorized and guest users are not
            allowed to create views
        '403':
          description: If the user is not allowed to create views
        '400':
          description: Bad request due to invalid parameters
    delete:
      summary: Delete Multiple Views
      operationId: bulkDeleteView
      description: Deletes multiple Views
      tags: 
        - Views
      parameters:
        - name: id
          in: query
          required: true
          description: Comma-separated list of IDs to delete
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              example:
                message: Deleted 5 graphs
        '401':
          description: If the user is not authorized and guest users are not
            allowed to delete included resources
        '403':
          description: If the user is not allowed to delete included resources
        '400':
          description: Bad request due to invalid parameters
  /views/{id}:
    get:
      summary: Get View by ID
      operationId: getView
      tags: 
        - Views
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/standardID'
        - name: tags
          in: query
          required: false
          description: If set include related tags with id, name and description
          schema:
            type: boolean
        - name: subscription
          in: query
          required: false
          description: If set include the subscription
          schema:
            type: boolean
        - name: group
          in: query
          required: false
          description: If set include the subscription and its group
          schema:
            type: boolean
        - name: study_areas
          in: query
          required: false
          description: If set include the subscription and its study areas
          schema:
            type: boolean
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/View'
                  - type: object
                    properties:
                      Tags:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              $ref: '#/components/schemas/standardID'
                            name:
                              type: string
                              maxLength: 50
                            description:
                              type: string
                              maxLength: 200
                      DataRequest:
                        type: object
                        properties:
                          id:
                            $ref: '#/components/schemas/standardID'
                          name:
                            type: string
                            maxLength: 100
                      group:
                        $ref: '#/components/schemas/SubscriptionGroup'
                      StudyAreas:
                        type: array
                        items:
                          $ref: '#/components/schemas/StudyArea'

              example:
                id: 1
                name: Example View
                description: Example View Description
                screenShot: Base 64 image data...
                settings: {}
                subscriptionId: 1
                creatorId: 2
                createdAt: 2024-02-22T13:36:17.315Z
                updatedAt: 2024-02-22T13:36:17.315Z
                Tags:
                  - id: 1
                    name: Tag 1
                    description: Tag 1 description
                DataRequest:
                  id: 1
                  name: Subscription Name
                group:
                  id: 1
                  name: Group 1
                  description: Group 1 description
                  createdAt: 2024-02-22T13:36:17.315Z
                  updatedAt: 2024-02-22T13:36:17.315Z
                StudyAreas: []

        '401':
          description: If the user is not authorized and guest users are not
            allowed to view included resources
        '403':
          description: If the user is not allowed to view included resources
        '400':
          description: Bad request due to invalid parameters
        '404':
          description: View not found
    put:
      summary: Update a View
      operationId: updateView
      description: |-
        Updates a View
        - The payload is a partial View object
        - Payload properties like `id`, `subscriptionId`, `createdAt` and
          `updatedAt` are ignored.
        - If `Tags` array is provided, it can contain partial tag objects but
          we only need/use their `id` property.
      tags: 
        - Views
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/standardID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                  example: View display name
                description:
                  type: string
                  maxLength: 500
                  example: View description
                screenShot:
                  type: string
                  example: base64imageData...
                Tags:
                  type: array
                  items:
                    $ref: '#/components/schemas/Tag'
                settings:
                  type: object
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/View'
        '401':
          description: If the user is not authorized and guest users are not
            allowed to update included resources
        '403':
          description: If the user is not allowed to update included resources
        '400':
          description: Bad request due to invalid parameters
        '500':
          description: Updating graph failed
    delete:
      summary: Delete a View
      operationId: deleteView
      description: Deletes a View
      tags: 
        - Views
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/standardID'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/View'
        '401':
          description: If the user is not authorized and guest users are not
            allowed to delete included resources
        '403':
          description: If the user is not allowed to delete included resources
        '400':
          description: Deleting failed
        '404':
          description: No View with this ID found
  /views/{id}/screenshot:
    get:
      summary: Get View Screenshot
      description: View have thumbnail images stored as base64 image data. If
        such data exists then convert it to an image file and serve it with
        proper caching headers. Otherwise serve the default view icon file.
      operationId: getViewScreenshot
      tags: 
        - Views
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/standardID'
      responses:
        '200':
          description: successful operation
          headers:
            Cache-Control:
              description: Disable caching
              schema:
                type: string
                example: max-age=31536000, no-cache
            Vary:
              description: Control caching using ETag
              schema:
                type: string
                example: Origin, ETag
            ETag:
              description: ETag for this image
              schema:
                type: string
            Content-Type:
              description: Image mime type
              schema:
                type: string
            Content-Length:
              description: Image byte size
              schema:
                type: integer
          content:
            image/png:
              schema:
                type: string
              example: BINARY IMAGE DATA

        '304':
          description: If the client provides a `if-none-match` header matching
            the current ETag of the image
        '400':
          description: Invalid base64 image data
        '401':
          description: If the user is not authorized and guest users are not
            allowed to view included resources
        '403':
          description: If the user is not allowed to view included resources
        '404':
          description: View not found
  /views/{id}/request-linelevel-data:
    post:
      summary: Request Linelevel Data
      description: This is not fully implemented yet. It will either change, or
        will be removed in the future
      deprecated: true
      operationId: requestLinelevelData
      tags: 
        - Views
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/standardID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                subscription:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                view:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                type:
                  type: string
                  enum:
                    - required
                    - preferred
                    - optional
                reason:
                  type: string
                user:
                  type: object
                  properties:
                    username:
                      type: string
                dataElements:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      need:
                        type: string
      responses:
        '200':
          description: Returns empty response on success
        '401':
          description: If the user is not authorized and guest users are not
            allowed to request linelevel data
        '403':
          description: If the user is not allowed to request linelevel data
        '400':
          description: Bad request due to invalid parameters

  # Subscriptions --------------------------------------------------------------
  /requests/{subscription-id}/views:
    get:
      summary: Get Views
      description: Get Views of the given subscription. Note that only views that
        the current user is allowed to see (if any) are included
      operationId: getSubscriptionViews
      tags:
        - Subscriptions
      parameters:
        - name: subscription-id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/standardID'
        # - $ref: '#/components/parameters/where'
        - $ref: '#/components/parameters/attributes'
        - $ref: '#/components/parameters/pick'
        - $ref: '#/components/parameters/omit'
        - $ref: '#/components/parameters/include'
        - $ref: '#/components/parameters/order'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/View'
        '401':
          description: If the user is not authorized and guest users are not
            allowed to view included resources
        '403':
          description: If the user is not allowed to view included resources
...
components:
  parameters:
    where:
      name: where
      in: query
      schema:
        type: string
      description: |-
        Can contain zero or more comma-separated where conditions.
        
        #### *Examples*:
        - ***name:value** pairs - `name:John`*
        - ***name:operator:value** - `age:lt:10`*
        - ***multiple conditions** - `name:John,age:gt:5`*
      required: false
      allowEmptyValue: true
    attributes:
      name: attributes
      in: query
      schema:
        type: string
      description: |-
        Comma-separated list of model attributes to include in the response

        ***Example**: `id, name, age`*
      required: false
      allowEmptyValue: true
    pick:
      name: pick
      in: query
      schema:
        type: string
      description: |-
        Comma-separated list of model attributes to include in the response.
        
        **NOTE:** This parameter is ignored when the `attributes` parameter is used

        ***Example**: `id, name, age`*
      required: false
      allowEmptyValue: true
    omit:
      name: omit
      in: query
      schema:
        type: string
      description: |-
        Comma-separated list of model attributes to exclude from the response.
        
        **NOTE:** This parameter is ignored when the `attributes` parameter is used

        ***Example**: `id, name, age`*
      required: false
      allowEmptyValue: true
...
    order:
      name: order
      in: query
      schema: 
        type: string
      description: |-
        Sorting instruction(s)

        ***Example**: `name:ask,age:desc`*
      required: false
      allowEmptyValue: true
...
    offset:
      name: offset
      in: query
      description: How many records to skip
      required: false
      allowEmptyValue: true
      schema: 
        type: integer
  schemas:
    standardID:
      type: integer
      format: int64
      minimum: 1
    standardName:
      type: string
...
    autoTimestamp:
      type: string
      format: date-time
...
    Tag:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/standardID'
        name:
          $ref: '#/components/schemas/standardName'
        description:
          $ref: '#/components/schemas/standardDescription'
        creatorId:
          $ref: '#/components/schemas/standardID'
        createdAt:
          $ref: '#/components/schemas/autoTimestamp'
        updatedAt:
          $ref: '#/components/schemas/autoTimestamp'
...
    StudyArea:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/standardID'
        name:
          type: string
          maxLength: 100
        description:
          type: string
        creatorId:
          type: integer
          minimum: 1
          nullable: true
        createdAt:
          $ref: '#/components/schemas/autoTimestamp'
        updatedAt:
          $ref: '#/components/schemas/autoTimestamp'
...
    SubscriptionGroup:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/standardID'
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          $ref: '#/components/schemas/autoTimestamp'
        updatedAt:
          $ref: '#/components/schemas/autoTimestamp'
    View:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/standardID'
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
          nullable: true
        screenShot:
          type: string
          nullable: true
        settings:
          type: object
        subscriptionId:
          $ref: '#/components/schemas/standardID'
        creatorId:
          $ref: '#/components/schemas/standardID'
        createdAt:
          $ref: '#/components/schemas/autoTimestamp'
        updatedAt:
          $ref: '#/components/schemas/autoTimestamp'
...
  securitySchemes:
    session_id:
      type: apiKey
      name: sid
      in: cookie